revert adding profiles (#69)
